chore: refactor structure of soruce related form file
