+++ src/components/forms/connector/source/CreateSourceForm/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./CreateSourceForm";
+export * from "./CreateSourceForm";
