+++ src/components/forms/connector/source/index.ts
@@ -1,0 +1,7 @@
+import ConfigureSourceForm from "./ConfigureSourceForm";
+import type { ConfigureSourceFormProps } from "./ConfigureSourceForm";
+import CreateSourceForm from "./CreateSourceForm";
+import type { CreateSourceFormValues } from "./CreateSourceForm";
+
+export { ConfigureSourceForm, CreateSourceForm };
+export type { ConfigureSourceFormProps, CreateSourceFormValues };
