+++ src/components/forms/index.ts
@@ -2,10 +2,10 @@ import CreatePipelineForm from "./CreatePipelineForm";
 import type { StepNumberState } from "./CreatePipelineForm/CreatePipelineForm";
 
 import CreateDestinationForm from "./CreateDestinationForm";
-import CreateSourceForm from "./CreateSourceForm";
+import CreateSourceForm from "./connector/source/CreateSourceForm";
 import OnboardingForm from "./OnboardingForm";
 import CreateModelForm from "./CreateModelForm";
-import ConfigureSourceForm from "./ConfigureSourceForm";
+import { ConfigureSourceForm } from "./connector";
 import ConfigureDestinationForm from "./ConfigureDestinationForm";
 import ConfigureModelForm from "./ConfigureModelForm";
 import ConfigureModelInstanceForm from "./ConfigureModelInstanceForm";
